oddecho
ECHO! Echo! Ech...

You love shouting in caves to hear your words echoed back at you. Unfortunately, as a hard-working software engineer you are not lucky enough to get out that often to shout in caves. Instead, you would like to implement a program that serves as a replacement for a cave.

Every now and then, you want to input a few words into the program and have them echoed back to you. However, as is well known, if you shout too quickly in a cave, the echo might cause interference with the new words you say. More specifically, every other word you say will interfere with the echo of your previous word. Thus only the first, third, fifth, and so on, words will actually produce an echo.

Your task is to write a program that simulates this behavior.

source: https://open.kattis.com/problems/oddecho