monopoly
Doris is playing monopoly with her friends, and it is now her turn to roll the dice. In standard monopoly you roll two 6-sided dice, and then walk the same number of steps as the sum of the two dice. From her current position, she will end up on her opponents’ hotels if the sum is equal to one of the integers A(1),...,A(N). What is the probability that she will end up on one of her opponent’ hotels during her current turn? You do not have to consider the rule about rerolling if both dice show the same number or any other rule from monopoly.

source: https://open.kattis.com/problems/monopol